Mark the repo as deprecated
Several people have pointed this out as a thing that should be used for new
projects, but after speaking to people on slack and in person, the central
instance of postcodeinfo is not being invested in any more and we should use OS
Places.

It was felt that for Moj's low usage levels, it's probably
more cost effective to use a commercial product.